Fix indexing issue.
When you use soft inheritance (with a proxy model), Django cannot
automatically downcast objects to the correct class we you iterate
over a queryset.

Hence, when reindexing documents, the wrong Metadataclass was used, and
the wrong PhaseConfig was selected.

To solve this issue, we extract the exact PhaseConfig version we need by
fetching it from the category which holds the exact Metadata subclass.